+++ changelog.md
@@ -1,12 +1,45 @@
 Changelog
 =========
 
-## v.2.5.0 - xx Dez, 2015
+## v.2.5.0-0 - xx Dez, 2015
 
 Features
 
--  The parsers moved into the [redis-parser](https://github.com/NodeRedis/node-redis-parser) module and will be maintained in there from now on ([@BridgeAR](https://github.com/BridgeAR))
- -  Improve js parser speed significantly for big SUNION/SINTER/LRANGE/ZRANGE ([@BridgeAR](https://github.com/BridgeAR))
+-  The parsers moved into the [redis-parser](https://github.com/NodeRedis/node-redis-parser) module and will be maintained in there from now on
+ -  Improve js parser speed significantly for big SUNION/SINTER/LRANGE/ZRANGE
+-  Improve redis-url parsing to also accept the database-number and options as query parameters as suggested in the [IANA](http://www.iana.org/assignments/uri-schemes/prov/redis)
+-  Added a `retry_unfulfilled_commands` option
+ -  Setting this to 'true' results in retrying all commands that were not fulfilled on a connection loss after the reconnect. Use with caution
+-  Added a `db` option to select the database while connecting (this is [not recommended](https://groups.google.com/forum/#!topic/redis-db/vS5wX8X4Cjg))
+-  Added a `password` option as alias for auth_pass
+-  The client.server_info is from now on updated while using the info command
+
+Bugfixes
+
+-  Fixed explicit undefined as a command callback in a multi context
+-  Fixed hmset failing to detect the first key as buffer or date if the key is of that type
+-  Fixed do not run toString on an array argument and throw a "invalid data" error instead
+ -  This is not considered as breaking change, as this is likely a error in your code and if you want to have such a behavior you should handle this beforehand
+ -  The same applies to Map / Set and individual Object types
+-  Fixed redis url not accepting the protocol being omitted or protocols other than the redis protocol for convienence
+-  Fixed parsing the db keyspace even if the first database does not begin with a zero
+-  Fixed handling of errors occuring while receiving pub sub messages
+
+Deprecations
+
+-  Using any command with a argument being set to null or undefined is deprecated
+ -  From v.3.0.0 on using a command with such an argument will return an error instead
+ -  If you want to keep the old behavior please use a precheck in your code that converts the arguments to a string.
+ -  Using SET or SETEX with a undefined or null value will from now on also result in converting the value to "null" / "undefined" to have a consistent behavior. This is not considered as breaking change, as it returned an error earlier.
+-  Using .end(flush) without the flush parameter deprecated and the flush parameter should explicitly be used
+ -  From v.3.0.0 on using .end without flush will result in an error
+ -  Using .end without flush means that any command that did not yet return is going to silently fail. Therefor this is considered harmfull and you should explicitly silence such errors if you are sure you want this
+-  Depending on the return value of a command to detect the backpressure is deprecated
+ -  From version 3.0.0 on node_redis might not return true / false as a return value anymore. Please rely on client.should_buffer instead
+-  The socket_nodelay option is deprecated and will be removed in v.3.0.0
+ -  If you want to buffer commands you should use [.batch or .multi](./README.md) instead. This is necessary to reduce the amount of different options and this is very likely reducing your throughput if set to false.
+ -  If you are sure you want to activate the NAGLE algorithm you can still activate it by using client.stream.setNoDelay(false)
+-  Redis < v. 2.6.11 is not supported anymore and will not work in all cases. Please update to a newer redis version
 
 ## v.2.4.2 - 27 Nov, 2015